+++ app/javascript/bali/index.js
@@ -2,8 +2,10 @@ import GoogleMapsLoader from './utils/google-maps-loader'
 import * as form from './utils/form'
 import * as formatters from './utils/formatters'
 import * as time from './utils/time'
+import * as domHelpers from './utils/domHelpers'
 
-export { GoogleMapsLoader, form, formatters, time }
+export { GoogleMapsLoader, form, formatters, time, domHelpers }
+export { AvatarController } from '../../components/bali/avatar'
 export { CarouselController } from '../../components/bali/carousel'
 export { ChartController } from '../../components/bali/chart'
 export { DrawerController } from '../../components/bali/drawer'