+++ spec/bali/components/icon_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Bali::Icon::Component, type: :component do
 
   before { @options = {} }
 
-  subject { rendered_component }
+  subject { page }
 
   it 'renders' do
     render_inline(component)